requirements.txt
torch==1.4.0
torchvision==0.5.0
tqdm==4.48.2
yacs==0.1.8
opencv-python==4.4.0.42
numpy==1.19.1
Pillow==7.2.0